# insis_api/swagger/swagger.yaml
---
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "Insis api"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
tags:
- name: "exam"
  description: "Insis student exam"
schemes:
- https
paths:
  /exam:
    post:
      tags:
      - "exam"
      summary: "Get all exams"
      description: ""
      operationId: "get_exam"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        description: "Insis user credentials"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Exam"
        404:
          description: "No exams found"
      x-swagger-router-controller: "insis_api.controllers.exam_controller"
  /exam/{examId}:
    post:
      tags:
      - "exam"
      summary: "Get exam by id"
      description: ""
      operationId: "get_exam_by_id"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "examId"
        in: "path"
        description: "ID of exam that needs to be fetched"
        required: true
        type: "integer"
        format: "int64"
      - name: "body"
        in: "body"
        description: "Insis user credentials"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Exam"
        404:
          description: "Exam not found"
      x-swagger-router-controller: "insis_api.controllers.exam_controller"
  /exam/{examId}/enroll:
    post:
      tags:
      - "exam"
      summary: "Enroll exam with id"
      description: ""
      operationId: "enroll_exam_by_id"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "examId"
        in: "path"
        description: "ID of exam to enroll"
        required: true
        type: "integer"
        format: "int64"
      - name: "body"
        in: "body"
        description: "Insis user credentials"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "Successful operation"
        400:
          description: "Unsuccessful operation"
        404:
          description: "Exam not found"
      x-swagger-router-controller: "insis_api.controllers.exam_controller"
  /exam/{examId}/disenroll:
    post:
      tags:
      - "exam"
      summary: "Disenroll exam with id"
      description: ""
      operationId: "disenroll_exam_by_id"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "examId"
        in: "path"
        description: "ID of exam to disenroll"
        required: true
        type: "integer"
        format: "int64"
      - name: "body"
        in: "body"
        description: "Insis user credentials"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "Successful operation"
        400:
          description: "Unsuccessful operation"
        404:
          description: "Exam not found"
      x-swagger-router-controller: "insis_api.controllers.exam_controller"
  /exam/enrolled:
    post:
      tags:
      - "exam"
      summary: "Get enrolled exams"
      description: ""
      operationId: "get_enrolled_exam"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        description: "Insis user credentials"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Exam"
        404:
          description: "No exams found"
      x-swagger-router-controller: "insis_api.controllers.exam_controller"
definitions:
  User:
    type: "object"
    required:
    - "username"
    - "password"
    properties:
      username:
        type: "string"
      password:
        type: "string"
  Exam:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      accessibility:
        type: "string"
      ident:
        type: "string"
      subject:
        type: "string"
      department:
        type: "string"
      date:
        type: "string"
        format: "date-time"
      room:
        type: "string"
      type:
        type: "string"
      teacher:
        type: "string"
      availability:
        type: "string"
      sign_in_from:
        type: "string"
        format: "date-time"
      sign_in_to:
        type: "string"
        format: "date-time"
      sign_out_to:
        type: "string"
        format: "date-time"
    example:
      date: "2000-01-23T04:56:07.000+00:00"
      teacher: "teacher"
      ident: "ident"
      subject: "subject"
      sign_out_to: "2000-01-23T04:56:07.000+00:00"
      id: 0
      availability: "availability"
      department: "department"
      type: "type"
      room: "room"
      sign_in_from: "2000-01-23T04:56:07.000+00:00"
      sign_in_to: "2000-01-23T04:56:07.000+00:00"